conversion: link-create.js to tsx
